+++ src/skill.h
@@ -150,7 +150,7 @@ class SkillLevel
 
         void train( int amount, bool skip_scaling = false );
         bool isRusting() const;
-        bool rust( bool charged_bio_mem, float character_rate );
+        bool rust( bool charged_bio_mem, int character_rate );
         void practice();
         bool can_train() const;
